Use built-in git instead of Git Automator extension
This commit updates the git instructions
To use the built-in VS Code git integration instead of Git Automator.

- Automatically stage all changes
  "git.enableSmartCommit": true,

- Automatically push & pull changes to the remote after committing
  "git.postCommitCommand": "sync"

Updates the settings in this project to dogfood the new defaults.

This feature could be improved further
By setting a default key-binding to Git: Commit All,
And/or by automatically committing on save or focus change
For the most SaaS-like experience.

Part of issue #30 along with a PR to foambubble/foam-template.